Fix more sprintf formats (or args)
git-svn-id: https://svn.isis.vanderbilt.edu/GMESRC/trunk@2810 b35830b6-564f-0410-ac5f-b387e50cb686

+++ SDK/BON/Common/ComponentObj.cpp
@@ -1003,12 +1003,12 @@ void CComponentObj::HandleError( Util::Exception* pEx )
 	strOut += "Exception message : " + pEx->getErrorMessage();
 	if ( pEx->getKind() == "MON::Exception" ) {
 		char chBuffer[ 100 ];
-		sprintf_s( chBuffer, "%x", ( (MON::Exception*) pEx)->getHResult() );
+		sprintf_s( chBuffer, "%x", (unsigned int)( (MON::Exception*) pEx)->getHResult() );
 		strOut += "\nException hresult : " + std::string( chBuffer );
 	}
 	else if ( pEx->getKind() == "BON::Exception" ) {
 		char chBuffer[ 100 ];
-		sprintf_s( chBuffer, "%x", ( (BON::Exception*) pEx)->getHResult() );
+		sprintf_s( chBuffer, "%x", (unsigned int)( (BON::Exception*) pEx)->getHResult() );
 		strOut += "\nException hresult : " + std::string( chBuffer );
 	}
 	AfxMessageBox( CString(strOut.c_str()) );
